Fixed bug #149 - SpectrumGroup setting in sidebar on load now works.
Problem was that notifiers on AbstractWrapperObject were always executed last,
so spectra in sidebar were refreshed before the SpectrumGroup was in the sidebar.
Solution was to replace the AbstractWrapperObject notifiers with individual class notifiers.

NB, there are currently no active AbstractWrapperObject notifiers left, 
but the machinery is in place for future use.

git-svn-id: https://svn.code.sf.net/p/ccpn/code/trunk@9843 8c04fef6-8de3-4ed0-b809-c157a10b9597