Modified the item _ListWidget so it can have elements in a sorted order
the order used is the same as the pids are listed in the project and
sidebar

this is not easy to do using an item based object!
(I wish we sued models)
this most probably removes the need for sortOnDrop as items can be
sorted by subclass ListWidgetItem which allows sorting on drops,
as well as addItem and any other approach...

some of this most probably needs to move into ListWidget and be used to
relace ListWidgetPair as it offers the same properties but in a more
flexible form?